requirements.txt
aiohttp>=3.9.0
websockets>=12.0
python-dotenv>=1.0.0
uvloop>=0.19.0; sys_platform != "win32"
